refactor: move AddTaskController class to their own production file
