+++ src/presentation/controller/index.ts
@@ -1,2 +1,3 @@
 export * from "./signUp-controller"
 export * from "./login-controller"
+export * from "./add-task-controller"
