+++ tests/presentation/controllers/add-task-controller.spec.ts
@@ -1,20 +1,8 @@
 
 import { MissingParamError } from "../../../src/presentation/errors"
 import { badRequest } from "../../../src/presentation/helpers"
-import { IController, IHttpRequest, IHttpResponse } from "../../../src/presentation/protocols"
+import { AddTaskController } from '../../../src/presentation/controller'
 
-export class AddTaskController implements IController {
-
-    async handle(httpRequest: IHttpRequest): Promise<IHttpResponse>{
-        const paramsRequired = ["title", "content"]
-        for (const params of paramsRequired) {
-            if(!httpRequest.body[params]){
-                return badRequest(new MissingParamError(params))
-            }
-        }
-    }
-    
-}
 
 
 describe('AddTaskController', () => {
